terminado el crud de productos

+++ DAW/2ndo/Módulo 7 - Desarrollo web en entorno servidor/Proyecto-Laravel/app/Http/Controllers/ProductController.php
@@ -17,6 +17,7 @@ class ProductController extends Controller
         $products->price = $request->price;
         $products->stock = $request->stock;
         $products->description = $request->description;
+        $products->category = $request->category;
         $products->save();
 
         return response()->json([
@@ -31,23 +32,38 @@ class ProductController extends Controller
         $products = new Products();
         $products->product_category = $request->category;
 
-        return Products::where('category', '=', $products->product_category,)->get();
+        return Products::where('category', '=', $products->product_category)->get();
+
     }
 
-    public function uploadProduct()
+    public function uploadProduct(Request $request)
     {
-        
+        $products = new Products();
+        $products->id = $request->id;
+        $products->product_name = $request->product_name;
+        $products->price = $request->price;
+        $products->stock = $request->stock;
+        $products->description = $request->description;
+        $products->category = $request->category;
+        $products = Products::find($products->id);
+        $products->update([
+            'product_name' => $request->product_name,
+            'price' => $request->price,
+            'stock' => $request->stock,
+            'description' => $request->description,
+            'category' => $request->category,
+
+        ]);
+        return response()->json([
+            "status" => 1,
+            "msg" => "¡El producto $products->product_name ha sido modificado correctamente!",
+        ]);
     }
-
-
 
     public function deleteProduct(Request $request)
     {
         $products = new Products();
         $products->id = $request->id;
-        if ($products = Products::find($products)) {
-            return Products::destroy($products);
-        }
-        return 'no va >:(';
+        if ($products = Products::find($products)) return Products::destroy($products);
     }
 }
